chore: TreeProvider and useTableTreeContext
