+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Tree/TreeRoot.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import TreeNodes from './TreeNodes';
-import { useTableTreeContext } from './TreeProvider';
+import { useTableTreeContext } from './TreeProvider/useTableTreeContext';
 
 const TreeRoot: FC = () => {
     const { nodeMap } = useTableTreeContext();
